style: rename simple chunker module
