+++ crates/ai/src/chunk/simple.rs
